pyproject.toml
[project]
name = "digital-fte"
version = "0.2.0"
description = "Digital FTE - Personal AI Employee (Silver Tier)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core dependencies
    "watchdog>=6.0.0",
    "fastmcp>=2.14.4",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    # Gmail integration
    "google-api-python-client>=2.192.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    # WhatsApp watcher (Playwright)
    "playwright>=1.40.0",
    "rich>=14.3.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
